requirements_tests.txt
aiohttp
coverage
gcsfs
geopy
gitpython
hypothesis-numpy
intake-esm
numba
pytest
scikit-learn
